requirements.txt
asgiref==3.4.1
Django==4.1.7
django-ckeditor==6.1.0
django-crispy-forms==1.13.0
django-js-asset==1.2.2
gunicorn==20.1.0
Pillow==9.4.0
psycopg2-binary==2.9.1
pytz==2021.3
sqlparse==0.4.2
django-modeltranslation==0.18.9
typing-extensions==4.5.0